Removed TaikoInputBlockingMod, and instead moved code into TaikoModSingleTap
